+++ CHANGELOG.md
@@ -6,5 +6,7 @@
 
 ### Fixed
 
+- A* tie-breaker preferred nodes farther from the target among equal F-costs by @bsommerfeld
+
 ### Removed